+++ java/src/main/java/cucumber/runtime/java/JavaStepDefinition.java
@@ -1,10 +1,8 @@
 package cucumber.runtime.java;
 
-import cucumber.annotation.Pending;
 import cucumber.runtime.CucumberException;
 import cucumber.runtime.JdkPatternArgumentMatcher;
 import cucumber.runtime.ParameterType;
-import cucumber.runtime.PendingException;
 import cucumber.runtime.StepDefinition;
 import gherkin.I18n;
 import gherkin.formatter.Argument;
@@ -33,9 +31,6 @@ public class JavaStepDefinition implements StepDefinition {
     }
 
     public void execute(I18n i18n, Object[] args) throws Throwable {
-        if (method.isAnnotationPresent(Pending.class)) {
-            throw new PendingException(method.getAnnotation(Pending.class).value());
-        }
         Class<?> clazz = method.getDeclaringClass();
         Object target = objectFactory.getInstance(clazz);
         try {